+++ src/PackageRegistryService/Pages/Handlers/PackageHandlers.cs
@@ -46,6 +46,7 @@ namespace PackageRegistryService.Pages.Handlers
                     packageReleaseNotes: package.ReleaseNotes ?? "",
                     packageProgrammingLanguage: package.ProgrammingLanguage,
                     packageAuthors: (package.Authors ?? []).ToArray(),
+                    packageCLIArguments: (package.CLIArguments ?? []).ToArray(),
                     versionTable: Components.PackageAvailableVersion.RenderVersionTable(packages),
                     downloads: downloads?.Downloads ?? 0
                 )
@@ -89,6 +90,7 @@ namespace PackageRegistryService.Pages.Handlers
                     packageReleaseNotes: latestPackage.ReleaseNotes ?? "",
                     packageProgrammingLanguage: latestPackage.ProgrammingLanguage,
                     packageAuthors: (latestPackage.Authors ?? []).ToArray(),
+                    packageCLIArguments: (latestPackage.CLIArguments ?? []).ToArray(),
                     versionTable: Components.PackageAvailableVersion.RenderVersionTable(packages),
                     downloads: downloads?.Downloads ?? 0
                 )